Broaden title and define "confirmed case"
I changed the tab and page title from "Coronavirus US" and "Coronavirus Cases in the US," respectively, to "2020 Coronavirus Pandemic in the United States." It's important to mention the year, 2020, in which the virus first spread in the nation. I also used "Pandemic" in lieu of "Cases" because the map now features county-wide statistics for confirmed cases, deaths, and the population. Finally, below the link is the WHO's official definition of "confirmed case," meant to clarify the counts given are tests satisfactorily performed by medical laboratories.